Add files via upload
NYS has adapted SAS code from the Colorado EPHT program, which gave incorrect numbers of sampling locations for HAA5's and TTHM's.  The Colorado code used a lot of PROC SQL, but we have favored PROC MEANS instead to calculate results.  It requires and input file consisting of standardized individual sample data.  The output is a CSV file which can be used as input for the XML generator.